d_a_npc_gk1: createInit switch-guard 93.7->98.0% + init_GK1_0 u8 (100%), add Nonmatching markers to 6 sub-100 fns

+++ include/d/actor/d_a_npc_gk1.h
@@ -21,7 +21,7 @@ public:
     void _nodeCB_Head(J3DNode*, J3DModel*);
     void _nodeCB_Neck(J3DNode*, J3DModel*);
     void _nodeCB_BackBone(J3DNode*, J3DModel*);
-    s32 init_GK1_0();
+    u8 init_GK1_0();
     u8 createInit();
     void play_animation();
     void setMtx(bool);
